cleaning code

+++ src/controllers/reviewControllers.js
@@ -55,7 +55,7 @@ exports.getReviewsByUserId = async (req, res) => {
 
 exports.getReviewsByMuseum = async (req, res) => {
   const museumId = req.params.museumId;
-
+  /*
   const [review, reviewdata] = await sequelize.query(
     `
   SELECT review.id, review.comment, review.grade, review.fk_museum_id AS museum, review.fk_user_id AS user_ID 
@@ -70,13 +70,14 @@ exports.getReviewsByMuseum = async (req, res) => {
   if (!review) {
     throw new NotFoundError("That museum does not have any reviews");
   }
-
-  const [results, metadata] = await sequelize.query(
+*/
+  const [review, metadata] = await sequelize.query(
     `
   SELECT  review.id, review.comment, review.grade, review.fk_museum_id AS museum, review.fk_user_id AS user_ID 
   FROM review
   JOIN museum ON museum.id = review.fk_museum_id
   WHERE review.fk_museum_id = $museumId
+  ORDER BY review.grade DESC
 
     `,
     {
@@ -87,7 +88,7 @@ exports.getReviewsByMuseum = async (req, res) => {
   if (!review || review.length == 0) {
     throw new NotFoundError("That museum does not have any reviews");
   }
-  return res.json(results);
+  return res.json(review);
 };
 
 exports.createNewReview = async (req, res) => {
@@ -120,7 +121,7 @@ exports.createNewReview = async (req, res) => {
 
 exports.deleteReviewById = async (req, res) => {
   const reviewId = req.params.reviewId;
-  console.log("jwjwjjfjafjf");
+  //console.log("jwjwjjfjafjf");
   const [review, reviewMeta] = await sequelize.query(
     `
         SELECT * FROM review
@@ -135,17 +136,24 @@ exports.deleteReviewById = async (req, res) => {
   if (!review) {
     throw new NotFoundError("This review does not exist.");
   }
-  await sequelize.query(
-    `
+  if (
+    req.user.role == userRoles.admin ||
+    req.user.userId == review.fk_user_id
+  ) {
+    await sequelize.query(
+      `
               DELETE FROM review
               WHERE review.id = $reviewId
               `,
-    {
-      bind: {
-        reviewId: reviewId,
-      },
-      types: QueryTypes.DELETE,
-    }
-  );
-  return res.sendStatus(204);
+      {
+        bind: {
+          reviewId: reviewId,
+        },
+        types: QueryTypes.DELETE,
+      }
+    );
+    return res.sendStatus(204);
+  } else {
+    throw new UnauthorizedError("You are not allowed to perform this action");
+  }
 };
